+++ src/kre/CanvasOGL.hpp
@@ -55,12 +55,12 @@ namespace KRE
 		void drawLine(const pointf& p1, const pointf& p2, const Color& color) const override;
 		// Draw filled polygon (i.e. triangle fan) using given color	
 		void drawPolygon(const std::vector<glm::vec2>& points, const Color& color) const override;
-		void drawSolidCircle(const point& centre, double radius, const Color& color) const override;
-		void drawSolidCircle(const point& centre, double radius, const std::vector<uint8_t>& color) const override;
-		void drawHollowCircle(const point& centre, double radius, const Color& color) const override;
-		void drawSolidCircle(const pointf& centre, double radius, const Color& color) const override;
-		void drawSolidCircle(const pointf& centre, double radius, const std::vector<uint8_t>& color) const override;
-		void drawHollowCircle(const pointf& centre, double radius, const Color& color) const override;
+		void drawSolidCircle(const point& centre, float radius, const Color& color) const override;
+		void drawSolidCircle(const point& centre, float radius, const std::vector<uint8_t>& color) const override;
+		void drawHollowCircle(const point& centre, float radius, const Color& color) const override;
+		void drawSolidCircle(const pointf& centre, float radius, const Color& color) const override;
+		void drawSolidCircle(const pointf& centre, float radius, const std::vector<uint8_t>& color) const override;
+		void drawHollowCircle(const pointf& centre, float radius, const Color& color) const override;
 		void drawPoints(const std::vector<glm::vec2>& points, float radius, const Color& color=Color::colorWhite()) const override;
 
 		static CanvasPtr getInstance();
